Implement compilation with CMake (#15)
* Write CMake file for building

* Update `setup.py` to use CMake when building the wheel

* Only fetch boost if not found

* Fix conditional compilation of CUDA

* Add compilation for hip in cmake

* Substitute old `CLUEstering.py` file

* Fix paths of modules when installing the library

* Change extension of cuda binding file to `.cu`

* Don't use `.cu` files

* Small tweaks to CMake file and formatting

* Fix typos

* Remove pybind11 from toml

* Fix packaging and installation

* Link shared object to `CLUEstering/lib folder`

* Delete copy of `CLUEstering.py`

* Cleaner check for existence of modules
